requirements.txt
python-dotenv
langchain
langchain-openai
pydantic
openai
flask
